Refactoring: GridCreatingConditionTriangle::createGrid() is now separated to an independent file, to make the function easier to read and understand.
